+++ basic_cms/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """Django Basic CMS module."""
-VERSION = (0, 1, 1)
+VERSION = (0, 1, 2)
 __version__ = '.'.join(map(str, VERSION))
 __author__ = "YD Technology"
 __contact__ = "team@ydtechnology.com"