+++ basic_cms/migrations/0001_initial.py
@@ -1,180 +1,185 @@
 # -*- coding: utf-8 -*-
-
+import datetime
 from south.db import db
+from south.v2 import SchemaMigration
 from django.db import models
 
-from basic_cms import settings
-from basic_cms.models import *
+from basic_cms.settings import PAGE_USER_MODEL
 
-class Migration:
+
+class Migration(SchemaMigration):
 
     def forwards(self, orm):
-
         # Adding model 'Page'
-        pages_page = [
-               ('id', orm['pages.Page:id']),
-               ('author', orm['pages.Page:author']),
-               ('parent', orm['pages.Page:parent']),
-               ('creation_date', orm['pages.Page:creation_date']),
-               ('publication_date', orm['pages.Page:publication_date']),
-               ('publication_end_date', orm['pages.Page:publication_end_date']),
-               ('last_modification_date', orm['pages.Page:last_modification_date']),
-               ('status', orm['pages.Page:status']),
-               ('template', orm['pages.Page:template']),
-               ('delegate_to', orm['pages.Page:delegate_to']),
-               ('redirect_to_url', orm['pages.Page:redirect_to_url']),
-               ('redirect_to', orm['pages.Page:redirect_to']),
-               ('lft', orm['pages.Page:lft']),
-               ('rght', orm['pages.Page:rght']),
-               ('tree_id', orm['pages.Page:tree_id']),
-               ('level', orm['pages.Page:level'])
-               ]
-        if settings.PAGE_TAGGING:
-            pages_page.append(('tags', orm['pages.Page:tags']))
-        db.create_table('pages_page', pages_page)
-        db.send_create_signal('pages', ['Page'])
+        db.create_table(u'basic_cms_page', (
+            (u'id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('author', self.gf('django.db.models.fields.related.ForeignKey')(related_name='pages', to=orm[PAGE_USER_MODEL])),
+            ('parent', self.gf('django.db.models.fields.related.ForeignKey')(blank=True, related_name='children', null=True, to=orm['basic_cms.Page'])),
+            ('creation_date', self.gf('django.db.models.fields.DateTimeField')(default=datetime.datetime(2013, 9, 6, 0, 0))),
+            ('publication_date', self.gf('django.db.models.fields.DateTimeField')(null=True, blank=True)),
+            ('publication_end_date', self.gf('django.db.models.fields.DateTimeField')(null=True, blank=True)),
+            ('last_modification_date', self.gf('django.db.models.fields.DateTimeField')()),
+            ('status', self.gf('django.db.models.fields.IntegerField')(default=0)),
+            ('template', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('delegate_to', self.gf('django.db.models.fields.CharField')(max_length=100, null=True, blank=True)),
+            ('freeze_date', self.gf('django.db.models.fields.DateTimeField')(null=True, blank=True)),
+            ('redirect_to_url', self.gf('django.db.models.fields.CharField')(max_length=200, null=True, blank=True)),
+            ('redirect_to', self.gf('django.db.models.fields.related.ForeignKey')(blank=True, related_name='redirected_pages', null=True, to=orm['basic_cms.Page'])),
+            ('lft', self.gf('django.db.models.fields.PositiveIntegerField')(db_index=True)),
+            ('rght', self.gf('django.db.models.fields.PositiveIntegerField')(db_index=True)),
+            ('tree_id', self.gf('django.db.models.fields.PositiveIntegerField')(db_index=True)),
+            ('level', self.gf('django.db.models.fields.PositiveIntegerField')(db_index=True)),
+        ))
+        db.send_create_signal(u'basic_cms', ['Page'])
+
+        # Adding M2M table for field sites on 'Page'
+        m2m_table_name = db.shorten_name(u'basic_cms_page_sites')
+        db.create_table(m2m_table_name, (
+            ('id', models.AutoField(verbose_name='ID', primary_key=True, auto_created=True)),
+            ('page', models.ForeignKey(orm[u'basic_cms.page'], null=False)),
+            ('site', models.ForeignKey(orm[u'sites.site'], null=False))
+        ))
+        db.create_unique(m2m_table_name, ['page_id', 'site_id'])
 
         # Adding model 'Content'
-        db.create_table('pages_content', (
-            ('id', orm['pages.Content:id']),
-            ('language', orm['pages.Content:language']),
-            ('body', orm['pages.Content:body']),
-            ('type', orm['pages.Content:type']),
-            ('page', orm['pages.Content:page']),
-            ('creation_date', orm['pages.Content:creation_date']),
+        db.create_table(u'basic_cms_content', (
+            (u'id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('language', self.gf('django.db.models.fields.CharField')(max_length=5)),
+            ('body', self.gf('django.db.models.fields.TextField')()),
+            ('type', self.gf('django.db.models.fields.CharField')(max_length=100, db_index=True)),
+            ('page', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['basic_cms.Page'])),
+            ('creation_date', self.gf('django.db.models.fields.DateTimeField')(default=datetime.datetime(2013, 9, 6, 0, 0))),
         ))
-        db.send_create_signal('pages', ['Content'])
+        db.send_create_signal(u'basic_cms', ['Content'])
 
         # Adding model 'PageAlias'
-        db.create_table('pages_pagealias', (
-            ('id', orm['pages.PageAlias:id']),
-            ('page', orm['pages.PageAlias:page']),
-            ('url', orm['pages.PageAlias:url']),
+        db.create_table(u'basic_cms_pagealias', (
+            (u'id', self.gf('django.db.models.fields.AutoField')(primary_key=True)),
+            ('page', self.gf('django.db.models.fields.related.ForeignKey')(to=orm['basic_cms.Page'], null=True, blank=True)),
+            ('url', self.gf('django.db.models.fields.CharField')(unique=True, max_length=255)),
         ))
-        db.send_create_signal('pages', ['PageAlias'])
-
-        # Adding model 'PagePermission'
-        db.create_table('pages_pagepermission', (
-            ('id', orm['pages.PagePermission:id']),
-            ('page', orm['pages.PagePermission:page']),
-            ('user', orm['pages.PagePermission:user']),
-            ('type', orm['pages.PagePermission:type']),
-        ))
-        db.send_create_signal('pages', ['PagePermission'])
-
-        if settings.PAGE_USE_SITE_ID:
-            # Adding ManyToManyField 'Page.sites'
-            db.create_table('pages_page_sites', (
-                ('id', models.AutoField(verbose_name='ID', primary_key=True, auto_created=True)),
-                ('page', models.ForeignKey(orm.Page, null=False)),
-                ('site', models.ForeignKey(orm['sites.Site'], null=False))
-            ))
-
+        db.send_create_signal(u'basic_cms', ['PageAlias'])
 
 
     def backwards(self, orm):
-
         # Deleting model 'Page'
-        db.delete_table('pages_page')
+        db.delete_table(u'basic_cms_page')
+
+        # Removing M2M table for field sites on 'Page'
+        db.delete_table(db.shorten_name(u'basic_cms_page_sites'))
 
         # Deleting model 'Content'
-        db.delete_table('pages_content')
+        db.delete_table(u'basic_cms_content')
 
         # Deleting model 'PageAlias'
-        db.delete_table('pages_pagealias')
-
-        # Deleting model 'PagePermission'
-        db.delete_table('pages_pagepermission')
-
-        if settings.PAGE_USE_SITE_ID:
-            # Dropping ManyToManyField 'Page.sites'
-            db.delete_table('pages_page_sites')
+        db.delete_table(u'basic_cms_pagealias')
 
 
-    page = {
-            'author': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']"}),
-            'creation_date': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+    models = {
+        u'accounts.account': {
+            'Meta': {'object_name': 'Account'},
+            'date_joined': ('django.db.models.fields.DateTimeField', [], {'auto_now_add': 'True', 'blank': 'True'}),
+            'email': ('django.db.models.fields.EmailField', [], {'unique': 'True', 'max_length': '110'}),
+            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'}),
+            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['auth.Group']", 'symmetrical': 'False', 'blank': 'True'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True'}),
+            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_university_admin': ('django.db.models.fields.BooleanField', [], {'default': 'False'}),
+            'is_university_professor': ('django.db.models.fields.BooleanField', [], {'default': 'False', 'db_index': 'True'}),
+            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
+            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '255', 'blank': 'True'}),
+            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
+            'token': ('django.db.models.fields.CharField', [], {'max_length': '40', 'null': 'True', 'blank': 'True'}),
+            'university': ('django.db.models.fields.related.ForeignKey', [], {'to': u"orm['universities.University']", 'null': 'True', 'blank': 'True'}),
+            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['auth.Permission']", 'symmetrical': 'False', 'blank': 'True'})
+        },
+        u'auth.group': {
+            'Meta': {'object_name': 'Group'},
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '80'}),
+            'permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': u"orm['auth.Permission']", 'symmetrical': 'False', 'blank': 'True'})
+        },
+        u'auth.permission': {
+            'Meta': {'ordering': "(u'content_type__app_label', u'content_type__model', u'codename')", 'unique_together': "((u'content_type', u'codename'),)", 'object_name': 'Permission'},
+            'codename': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
+            'content_type': ('django.db.models.fields.related.ForeignKey', [], {'to': u"orm['contenttypes.ContentType']"}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        },
+        u'basic_cms.content': {
+            'Meta': {'object_name': 'Content'},
+            'body': ('django.db.models.fields.TextField', [], {}),
+            'creation_date': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime(2013, 9, 6, 0, 0)'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'language': ('django.db.models.fields.CharField', [], {'max_length': '5'}),
+            'page': ('django.db.models.fields.related.ForeignKey', [], {'to': u"orm['basic_cms.Page']"}),
+            'type': ('django.db.models.fields.CharField', [], {'max_length': '100', 'db_index': 'True'})
+        },
+        u'basic_cms.page': {
+            'Meta': {'ordering': "['tree_id', 'lft']", 'object_name': 'Page'},
+            'author': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "'pages'", 'to': u"orm['accounts.Account']"}),
+            'creation_date': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime(2013, 9, 6, 0, 0)'}),
             'delegate_to': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'freeze_date': ('django.db.models.fields.DateTimeField', [], {'null': 'True', 'blank': 'True'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'last_modification_date': ('django.db.models.fields.DateTimeField', [], {}),
             'level': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
             'lft': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
-            'parent': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "'children'", 'null': 'True', 'to': "orm['pages.Page']"}),
+            'parent': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "'children'", 'null': 'True', 'to': u"orm['basic_cms.Page']"}),
             'publication_date': ('django.db.models.fields.DateTimeField', [], {'null': 'True', 'blank': 'True'}),
             'publication_end_date': ('django.db.models.fields.DateTimeField', [], {'null': 'True', 'blank': 'True'}),
-            'redirect_to': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "'redirected_pages'", 'null': 'True', 'to': "orm['pages.Page']"}),
+            'redirect_to': ('django.db.models.fields.related.ForeignKey', [], {'blank': 'True', 'related_name': "'redirected_pages'", 'null': 'True', 'to': u"orm['basic_cms.Page']"}),
             'redirect_to_url': ('django.db.models.fields.CharField', [], {'max_length': '200', 'null': 'True', 'blank': 'True'}),
             'rght': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'}),
+            'sites': ('django.db.models.fields.related.ManyToManyField', [], {'default': '[1]', 'related_name': "'pages'", 'symmetrical': 'False', 'to': u"orm['sites.Site']"}),
             'status': ('django.db.models.fields.IntegerField', [], {'default': '0'}),
             'template': ('django.db.models.fields.CharField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
             'tree_id': ('django.db.models.fields.PositiveIntegerField', [], {'db_index': 'True'})
-            }
-    if settings.PAGE_TAGGING:
-        page['tags'] = ('tagging.fields.TagField', [], {'null': 'True'})
-    if settings.PAGE_USE_SITE_ID:
-        page['sites'] = ('django.db.models.fields.related.ManyToManyField', [], {'default': '[1]', 'to': "orm['sites.Site']"})
-
-    models = {
-        'auth.group': {
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '80'}),
-            'permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': "orm['auth.Permission']", 'blank': 'True'})
         },
-        'auth.permission': {
-            'Meta': {'unique_together': "(('content_type', 'codename'),)"},
-            'codename': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
-            'content_type': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['contenttypes.ContentType']"}),
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        u'basic_cms.pagealias': {
+            'Meta': {'object_name': 'PageAlias'},
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'page': ('django.db.models.fields.related.ForeignKey', [], {'to': u"orm['basic_cms.Page']", 'null': 'True', 'blank': 'True'}),
+            'url': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '255'})
         },
-        'auth.user': {
-            'date_joined': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
-            'email': ('django.db.models.fields.EmailField', [], {'max_length': '75', 'blank': 'True'}),
-            'first_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
-            'groups': ('django.db.models.fields.related.ManyToManyField', [], {'to': "orm['auth.Group']", 'blank': 'True'}),
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'is_active': ('django.db.models.fields.BooleanField', [], {'default': 'True', 'blank': 'True'}),
-            'is_staff': ('django.db.models.fields.BooleanField', [], {'default': 'False', 'blank': 'True'}),
-            'is_superuser': ('django.db.models.fields.BooleanField', [], {'default': 'False', 'blank': 'True'}),
-            'last_login': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
-            'last_name': ('django.db.models.fields.CharField', [], {'max_length': '30', 'blank': 'True'}),
-            'password': ('django.db.models.fields.CharField', [], {'max_length': '128'}),
-            'user_permissions': ('django.db.models.fields.related.ManyToManyField', [], {'to': "orm['auth.Permission']", 'blank': 'True'}),
-            'username': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '30'})
-        },
-        'contenttypes.contenttype': {
-            'Meta': {'unique_together': "(('app_label', 'model'),)", 'db_table': "'django_content_type'"},
+        u'contenttypes.contenttype': {
+            'Meta': {'ordering': "('name',)", 'unique_together': "(('app_label', 'model'),)", 'object_name': 'ContentType', 'db_table': "'django_content_type'"},
             'app_label': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'model': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
             'name': ('django.db.models.fields.CharField', [], {'max_length': '100'})
         },
-        'pages.content': {
-            'body': ('django.db.models.fields.TextField', [], {}),
-            'creation_date': ('django.db.models.fields.DateTimeField', [], {'default': 'datetime.datetime.now'}),
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'language': ('django.db.models.fields.CharField', [], {'max_length': '5'}),
-            'page': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['pages.Page']"}),
-            'type': ('django.db.models.fields.CharField', [], {'max_length': '100'})
-        },
-        'pages.page': page,
-        'pages.pagealias': {
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'page': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['pages.Page']", 'null': 'True', 'blank': 'True'}),
-            'url': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '255'})
-        },
-        'pages.pagepermission': {
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
-            'page': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['pages.Page']", 'null': 'True', 'blank': 'True'}),
-            'type': ('django.db.models.fields.IntegerField', [], {'default': '0'}),
-            'user': ('django.db.models.fields.related.ForeignKey', [], {'to': "orm['auth.User']"})
-        },
-        'sites.site': {
-            'Meta': {'db_table': "'django_site'"},
+        u'sites.site': {
+            'Meta': {'ordering': "('domain',)", 'object_name': 'Site', 'db_table': "'django_site'"},
             'domain': ('django.db.models.fields.CharField', [], {'max_length': '100'}),
-            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
             'name': ('django.db.models.fields.CharField', [], {'max_length': '50'})
+        },
+        u'taggit.tag': {
+            'Meta': {'object_name': 'Tag'},
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'unique': 'True', 'max_length': '100'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'unique': 'True', 'max_length': '100'})
+        },
+        u'taggit.taggeditem': {
+            'Meta': {'object_name': 'TaggedItem'},
+            'content_type': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "u'taggit_taggeditem_tagged_items'", 'to': u"orm['contenttypes.ContentType']"}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'object_id': ('django.db.models.fields.IntegerField', [], {'db_index': 'True'}),
+            'tag': ('django.db.models.fields.related.ForeignKey', [], {'related_name': "u'taggit_taggeditem_items'", 'to': u"orm['taggit.Tag']"})
+        },
+        u'universities.university': {
+            'Meta': {'object_name': 'University'},
+            'description': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'domain': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            u'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'info': ('django.db.models.fields.TextField', [], {'blank': 'True'}),
+            'logo': ('django.db.models.fields.files.ImageField', [], {'max_length': '100', 'null': 'True', 'blank': 'True'}),
+            'name': ('django.db.models.fields.CharField', [], {'max_length': '255'}),
+            'slug': ('django.db.models.fields.SlugField', [], {'max_length': '50'})
         }
     }
 
-    complete_apps = ['pages']
+    complete_apps = ['basic_cms']